feat: highlight current page in navbar
